+++ app/src/main/java/com/madaex/exchange/ui/buy/fragment/Buy2Fragment.java
@@ -223,7 +223,7 @@ public class Buy2Fragment extends BaseNetLazyFragment<DealPresenter> implements 
         mBuyrecyclerview.setHasFixedSize(true);
         mBuyAdapter.setItemClickListener(new BuyAdapter.OnItemClickListener() {
             @Override
-            public void onItemClick(List<Float> list) {
+            public void onItemClick(List<BigDecimal> list) {
                 if (market_type.equals("0")) {
                     mPrice.setText(list.get(0).toString());
                 } else {
@@ -244,7 +244,7 @@ public class Buy2Fragment extends BaseNetLazyFragment<DealPresenter> implements 
         mSellerrecyclerview.setHasFixedSize(true);
         mSellerAdapter.setItemClickListener(new SellerAdapter.OnItemClickListener() {
             @Override
-            public void onItemClick(List<Float> list) {
+            public void onItemClick(List<BigDecimal> list) {
                 if (market_type.equals("0")) {
                     mPrice.setText(list.get(0).toString());
                 } else {
@@ -307,22 +307,22 @@ public class Buy2Fragment extends BaseNetLazyFragment<DealPresenter> implements 
     }
 
 
-    private List<DepthDataBean> getBuyDepthList2(List<List<Float>> lists) {
+    private List<DepthDataBean> getBuyDepthList2(List<List<BigDecimal>> lists) {
         Log.v("depthList", lists.size() + "");
         List<DepthDataBean> depthList = new ArrayList<>();
         if (lists.size() >= 50) {
-            List<List<Float>> listList = lists.subList(0, 50);
+            List<List<BigDecimal>> listList = lists.subList(0, 50);
 
             double account = 0;
             for (int i = 0; i < listList.size(); i++) {
-                account = account + listList.get(i).get(0);
-                depthList.add(new DepthDataBean(listList.get(i).get(0),
-                        listList.get(i).get(1)));
+                account = account + listList.get(i).get(0).floatValue();
+                depthList.add(new DepthDataBean(listList.get(i).get(0).floatValue(),
+                        listList.get(i).get(1).floatValue()));
             }
         } else {
             for (int i = 0; i < lists.size(); i++) {
-                depthList.add(new DepthDataBean(lists.get(i).get(0),
-                        lists.get(i).get(1)));
+                depthList.add(new DepthDataBean(lists.get(i).get(0).floatValue(),
+                        lists.get(i).get(1).floatValue()));
             }
         }
 
@@ -338,30 +338,30 @@ public class Buy2Fragment extends BaseNetLazyFragment<DealPresenter> implements 
 
     }
 
-    private List<DepthDataBean> getBuyDepthList3(List<List<Float>> buy, List<List<Float>> sell) {
+    private List<DepthDataBean> getBuyDepthList3(List<List<BigDecimal>> buy, List<List<BigDecimal>> sell) {
         List<DepthDataBean> depthList = new ArrayList<>();
         if (buy.size() >= 50 && sell.size() >= 50) {
-            List<List<Float>> listList = buy.subList(0, 49);
-            List<List<Float>> listList2 = sell.subList(0, 49);
+            List<List<BigDecimal>> listList = buy.subList(0, 49);
+            List<List<BigDecimal>> listList2 = sell.subList(0, 49);
             double account = 0;
             double account2 = 0;
             if (type.equals(ConstantUrl.TRANS_TYPE_BUY)) {
                 for (int i = 0; i < listList.size(); i++) {
-                    account = account + listList.get(i).get(0);
-                    account2 = account2 + listList.get(i).get(0);
+                    account = account + listList.get(i).get(0).doubleValue();
+                    account2 = account2 + listList.get(i).get(0).doubleValue();
 
                 }
                 for (int j = 0; j < listList2.size(); j++) {
-                    account = account + listList2.get(j).get(0);
+                    account = account + listList2.get(j).get(0).doubleValue();
                 }
                 mBili.setText(bs(account2, account) + "%");
             } else {
                 for (int i = 0; i < listList.size(); i++) {
-                    account = account + listList.get(i).get(0);
+                    account = account + listList.get(i).get(0).doubleValue();
                 }
                 for (int j = 0; j < listList2.size(); j++) {
-                    account = account + listList2.get(j).get(0);
-                    account2 = account2 + listList2.get(j).get(0);
+                    account = account + listList2.get(j).get(0).doubleValue();
+                    account2 = account2 + listList2.get(j).get(0).doubleValue();
                 }
                 mBili.setText(bs(account2, account) + "%");
             }
@@ -371,21 +371,21 @@ public class Buy2Fragment extends BaseNetLazyFragment<DealPresenter> implements 
             double account2 = 0;
             if (type.equals(ConstantUrl.TRANS_TYPE_BUY)) {
                 for (int i = 0; i < buy.size(); i++) {
-                    account = account + buy.get(i).get(0);
-                    account2 = account2 + buy.get(i).get(0);
+                    account = account + buy.get(i).get(0).doubleValue();
+                    account2 = account2 + buy.get(i).get(0).doubleValue();
 
                 }
                 for (int j = 0; j < sell.size(); j++) {
-                    account = account + sell.get(j).get(0);
+                    account = account + sell.get(j).get(0).doubleValue();
                 }
                 mBili.setText(bs(account2, account) + "%");
             } else {
                 for (int i = 0; i < buy.size(); i++) {
-                    account = account + buy.get(i).get(0);
+                    account = account + buy.get(i).get(0).doubleValue();
                 }
                 for (int j = 0; j < sell.size(); j++) {
-                    account = account + sell.get(j).get(0);
-                    account2 = account2 + sell.get(j).get(0);
+                    account = account + sell.get(j).get(0).doubleValue();
+                    account2 = account2 + sell.get(j).get(0).doubleValue();
                 }
                 mBili.setText(bs(account2, account) + "%");
             }
@@ -394,19 +394,19 @@ public class Buy2Fragment extends BaseNetLazyFragment<DealPresenter> implements 
         return depthList;
     }
 
-    private List<DepthDataBean> getSellDepthList2(List<List<Float>> lists) {
+    private List<DepthDataBean> getSellDepthList2(List<List<BigDecimal>> lists) {
         List<DepthDataBean> depthList = new ArrayList<>();
         if (lists.size() >= 50) {
-            List<List<Float>> listList = lists.subList(0, 50);
+            List<List<BigDecimal>> listList = lists.subList(0, 50);
             Log.v("depthList", listList.size() + "");
             for (int i = 0; i < listList.size(); i++) {
-                depthList.add(new DepthDataBean(listList.get(i).get(0),
-                        listList.get(i).get(1)));
+                depthList.add(new DepthDataBean(listList.get(i).get(0).floatValue(),
+                        listList.get(i).get(1).floatValue()));
             }
         } else {
             for (int i = 0; i < lists.size(); i++) {
-                depthList.add(new DepthDataBean(lists.get(i).get(0),
-                        lists.get(i).get(1)));
+                depthList.add(new DepthDataBean(lists.get(i).get(0).floatValue(),
+                        lists.get(i).get(1).floatValue()));
             }
         }
 
@@ -664,7 +664,7 @@ public class Buy2Fragment extends BaseNetLazyFragment<DealPresenter> implements 
         mBuyrecyclerview.setHasFixedSize(true);
         mBuyAdapter.setItemClickListener(new BuyAdapter.OnItemClickListener() {
             @Override
-            public void onItemClick(List<Float> list) {
+            public void onItemClick(List<BigDecimal> list) {
                 if (market_type.equals("0")) {
                     mPrice.setText(list.get(0).toString());
                 } else {
